+++ src/counter.ts
@@ -1,6 +1,8 @@
 // Monitorado: counter.ts
 
 /** @hidden */
+const _ = require("lodash")
+/** @hidden */
 const logger = require("anyhow")
 /** @hidden */
 const moment = require("moment")
@@ -10,7 +12,7 @@ export interface CounterOptions {
     /** metric should expire in these amount of milliseconds if not ended. */
     expiresIn?: number
     /** Custom tag or text to be associated with that particular metric. */
-    tag?: string
+    tag?: string | number | boolean | Date
 }
 
 /** Represents a single metric counter with time, duratiomn and additional info. */
@@ -22,7 +24,13 @@ export class Counter {
 
         // Append tag?
         if (options.tag) {
-            this.tag = options.tag
+            if (_.isDate(options.tag)) {
+                this.tag = options.tag.toString()
+            } else if (!_.isObject(options.tag)) {
+                this.tag = options.tag
+            } else {
+                throw new Error("The tag should be a string, number, boolean or date.")
+            }
         }
 
         // Should the metric expire (value in milliseconds)?
@@ -51,7 +59,7 @@ export class Counter {
     /** Optional, will be true if the counter has expired. */
     expired?: boolean
     /** Optional tag or label with extra information about the counter. */
-    tag?: string
+    tag?: string | number | boolean | Date
     /** Optional data for the counter. */
     data?: any
     /** Optional error object or string in case counter ended but there was an error. */